HW3.py
Write to function
The name of the first function is prime_first.
The name of the second function is prime_second.
You must use these two function inside the for loop. Ensure that the for loop takes values between 0-1000.
Press the prime numbers between 500-100  on the screen with the prime_second function.